+++ packages/be-config/src/__tests__/configurationMocks/validAuthenticationConfig.json
