+++ packages/be-config/src/__tests__/configurationMocks/validMicrolcConfig.json
