+++ packages/be-config/src/apis/configurationApi.ts
@@ -1,7 +1,7 @@
 import {Configuration, configurationSchema} from '@mia-platform/core'
 import {DecoratedFastify, Handler} from '@mia-platform/custom-plugin-lib'
 
-import {readValidateConfiguration} from './configurationManager'
+import {readValidateConfiguration} from '../utils/configurationManager'
 
 const readPluginConfiguration = async(fastifyInstance: DecoratedFastify) => {
   // @ts-ignore
