+++ packages/be-config/src/constants.ts
@@ -14,8 +14,3 @@ export const CONFIGURATION_ENDPOINT: EndpointConfiguration = {
   METHOD: 'GET',
   PATH: `/api/${API_VERSION}/microlc/configuration`,
 }
-
-export const AUTHENTICATION_ENV = {
-  IS_AUTH_NECESSARY: 'MICROLC_AUTH_NECESSARY',
-  USER_INFO_URL: 'MICROLC_USER_INFO_URL',
-}
