+++ packages/be-config/src/index.ts
@@ -1,8 +1,8 @@
 import customService, {DecoratedFastify} from '@mia-platform/custom-plugin-lib'
 
 import {AUTHENTICATION_ENDPOINT, CONFIGURATION_ENDPOINT} from './constants'
-import {authenticationApiHandlerBuilder, authenticationApiSchema} from './authenticationApi'
-import {configurationApiHandlerBuilder, configurationApiSchema} from './configurationApi'
+import {authenticationApiHandlerBuilder, authenticationApiSchema} from './apis/authenticationApi'
+import {configurationApiHandlerBuilder, configurationApiSchema} from './apis/configurationApi'
 import {environmentVariablesSchema} from './environmentVariablesSchema'
 
 module.exports = customService(environmentVariablesSchema)(async function index(service: DecoratedFastify) {
