+++ packages/be-config/tsconfig.json
@@ -24,5 +24,10 @@
   },
   "include": [
     "src"
-  ]
+  ],
+  "baseUrl": "./src",
+  "paths": {
+    "@utils/*": ["./utils/*"],
+    "@apis/*": ["./apis/*"]
+  }
 }
